// src/lib/database.types.ts
export type Category = {
  id: number
  name: string
  created_at: string
}